requirements.txt
argparse==1.1.0
ftfy==6.1.1
grpcio==1.59.2
h5py==3.10.0
httpx==0.26.0
huggingface-hub==0.14.1
joblib==1.2.0
json==2.0.9
matplotlib==3.7.2
numpy==1.24.3
pandas==2.0.1
Pillow==9.5.0
regex==2023.5.5
sklearn==0.0
scipy==1.10.1
torch==2.0.1
torchvision==0.15.2
transformers==4.29.2
tokenizers==0.13.3
tqdm==4.65.0